Added a debugger file
